Add full mocking support for service and db tests
For the beta program, we need the application to be able to run unit
tests without breaking the state of the application for normal use.  For
example, when testing deleting records from the database, we do not want
the actual database to be modified.  Another concern is that the
application will be written against a specific server API, and that API
is actively changing.  It would be useful for the application to self
contain its own pseudo-service that can serve static requests based on
what the server API was at the time of writing the tests.  This also
eliminates a drop in internet connectivity causing unit tests to fail.

This change adds mocking support for both server communication and
database storage.  When the application is created, it checks a
parameter set by BetaManager to determine if the application should use
mock data sources and if so, initialize its state as necessary.  If the
beta manager is not present or that it is not set, normal data sources
are used instead.