Update how campus services are displayed
Instead of using a treeview or expandablelistview, the campus services
activity now displays a single category at a time.  Categories can be
clicked, launching a new instance of the activity to view that category.

services can be searched as usual, and the results of the search are
displayed in a new activity that displays the name of the link and the
path to that entry.